backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.16,<3.17
django-cors-headers>=4.9,<5.0
celery>=5.3,<6.0
redis>=5.0,<6.0
PyYAML>=6.0,<7.0
cryptography>=46.0,<47.0
uvicorn[standard]>=0.35,<0.39
whitenoise>=6.9,<7.0
